refactor (post) : post 도메인 DDD 구조 명확히 정의
