+++ src/main/java/com/backend/immilog/post/application/services/PostUploadService.java
@@ -6,8 +6,8 @@ import com.backend.immilog.post.domain.model.PostResource;
 import com.backend.immilog.post.domain.repositories.BulkInsertRepository;
 import com.backend.immilog.post.domain.repositories.PostRepository;
 import com.backend.immilog.post.exception.PostException;
-import com.backend.immilog.user.application.services.UserInformationService;
-import com.backend.immilog.user.domain.model.User;
+import com.backend.immilog.user.application.services.query.UserQueryService;
+import com.backend.immilog.user.domain.model.user.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -22,13 +22,14 @@ import static com.backend.immilog.post.domain.model.enums.PostType.POST;
 import static com.backend.immilog.post.domain.model.enums.ResourceType.ATTACHMENT;
 import static com.backend.immilog.post.domain.model.enums.ResourceType.TAG;
 import static com.backend.immilog.post.exception.PostErrorCode.FAILED_TO_SAVE_POST;
+import static com.backend.immilog.user.exception.UserErrorCode.USER_NOT_FOUND;
 
 @Slf4j
 @RequiredArgsConstructor
 @Service
 public class PostUploadService {
     private final PostRepository postRepository;
-    private final UserInformationService userInformationService;
+    private final UserQueryService userQueryService;
     private final BulkInsertRepository bulkInsertRepository;
 
     @Transactional
@@ -36,23 +37,11 @@ public class PostUploadService {
             Long userSeq,
             PostUploadCommand postUploadCommand
     ) {
-        User user = userInformationService.getUser(userSeq);
-        Post post = postRepository.saveEntity(
-                createPostEntity(userSeq, postUploadCommand, user)
-        );
+        User user = userQueryService.getUserById(userSeq)
+                .orElseThrow(() -> new PostException(USER_NOT_FOUND));
+        Post post = postRepository.save(Post.of(postUploadCommand, user));
         Long postSeq = post.seq();
         insertAllPostResources(postUploadCommand, postSeq);
-    }
-
-    private static Post createPostEntity(
-            Long userSeq,
-            PostUploadCommand postUploadCommand,
-            User user
-    ) {
-        return Post.of(
-                postUploadCommand,
-                user
-        );
     }
 
     private void insertAllPostResources(
@@ -66,13 +55,13 @@ public class PostUploadService {
         bulkInsertRepository.saveAll(
                 postResourceList,
                 """
-                INSERT INTO post_resource (
-                    post_seq,
-                    post_type,
-                    resource_type,
-                    content
-                ) VALUES (?, ?, ?, ?)
-                """,
+                        INSERT INTO post_resource (
+                            post_seq,
+                            post_type,
+                            resource_type,
+                            content
+                        ) VALUES (?, ?, ?, ?)
+                        """,
                 (ps, postResource) -> {
                     try {
                         ps.setLong(1, postResource.postSeq());
