+++ src/main/java/com/backend/immilog/post/domain/repositories/PostRepository.java
@@ -38,7 +38,7 @@ public interface PostRepository {
             Long postSeq
     );
 
-    Post saveEntity(
+    Post save(
             Post postEntity
     );
 }
