+++ src/main/java/com/backend/immilog/post/infrastructure/jdbc/CommentJdbcRepository.java
@@ -4,8 +4,8 @@ import com.backend.immilog.global.enums.UserRole;
 import com.backend.immilog.post.domain.model.enums.PostStatus;
 import com.backend.immilog.post.infrastructure.result.CommentEntityResult;
 import com.backend.immilog.user.application.result.UserInfoResult;
-import com.backend.immilog.user.domain.model.enums.UserCountry;
-import com.backend.immilog.user.domain.model.enums.UserStatus;
+import com.backend.immilog.user.domain.enums.UserCountry;
+import com.backend.immilog.user.domain.enums.UserStatus;
 import lombok.RequiredArgsConstructor;
 import org.springframework.jdbc.core.simple.JdbcClient;
 import org.springframework.stereotype.Repository;
