+++ src/main/java/com/backend/immilog/post/infrastructure/result/JobBoardEntityResult.java
@@ -4,7 +4,7 @@ import com.backend.immilog.post.application.result.JobBoardResult;
 import com.backend.immilog.post.domain.model.enums.*;
 import com.backend.immilog.post.infrastructure.jpa.InteractionUserEntity;
 import com.backend.immilog.post.infrastructure.jpa.PostResourceEntity;
-import com.backend.immilog.user.domain.model.enums.Industry;
+import com.backend.immilog.user.domain.enums.Industry;
 import lombok.Getter;
 import lombok.Setter;
